Completed integration test.
To run the test, run the script
projects/integration-test/main.py.
this script will rebuild the network-table
client and server, run the server, then
run the multiple clients. It
will count the total number of errors which occured
and print it out. The number of errors which occured
is passed as the return value from the client.