Implemented GETVALUE and SETVALUE
Values are stored in a std::map.
Add Reply.proto and GetValueReply.proto.
Renamed Message to Request.
Renamed SetKeyAction to SetKeyRequest.
Renamed GetKeyAction to GetKeyRequest.

Any protobuf with "Request" in its name
is meant to be sent from the client to the
server. Any with "Reply" in its name is
meant to be sent from the server to
the client.