NetworkTable now uses a tree structure
The table is made of of NetworkTable::Nodes which are a protobuf class.
A NetworkTable::Node contains a map of NetworkTable::Nodes which are its
children, and contains a value. Technically only leaves should have
values, but I did not implement this.

The values are now stored in a class called NetworkTable::Tree,
which contains two methods: GetNode and SetNode. When the tree
is first created, it contains a single NetworkTable::Node, the root.

I also created a new project called "viewtree" which can be used
to view the tree structure of the currently running NetworkTable.
The print function could probably be improved to look more like a tree,
but right now it gives you a general idea of what's currently in the
NetworkTable.

Switch from proto2 to proto3 protoc.
This is because proto3 has the "map", unlike proto2