Delete socket on client side after disconnect
If the server isn't running when the disconnect request is sent,
the server won't receive the disconnect message and will
later try to reconnect to the now abandoned socket.
To avoid this, also delete the socket on the client side.
This isn't the cleanest solution since the socket is now
being deleted by both the client and the server. However it's
better than blocking the client and waiting for the server (which may
not be running) to receive the disconnect request.