Fix race condition
Previously the code would check the "connected_" flag to see if the
ManageSocket thread had connected. However, sometimes the main thread
will time out while waiting for connected_ to be set, and then
immediately after connected_ will be set to true (race condition).
Fixed this by instead making the ManageSocket thread send a message to
the main thread when it either connects or fails to connect to the
server.