Added dummy satellite program
This program will sync two network tables connected by an ethernet
cable. Whenever a change is made to the local network table, these
changes are sent across the ethernet cable to be applied to the remote
network table.
Some of the source code had to be updated to allow the user to know if a
subscription update occured because of their own SetValues request, or
because some other process sent a SetValues request. Without this, an
infinite loop where the program would update the network table, and send
this change across the ethernet cable, and the program on the other side
would do the same.

To run this, the side that starts first has to call zmq bind and bind
to the ip address of their own ethernet port, and the other side then
has to call zmq connect, and connect to the remote ip address,
so the code has to be changed before running. This should be fixed
soon.