Added command line args for dummy-satellite
One side acts as a server, and the other a client. The code on both
sides is 99% the same, its just that zmq needs to call bind on the
server side, and connect on the client side. Command line args can now
be used to specify which side is client and which side is the server.