Added helper function to convert generic network table node to Nt.proto
Nt.proto is a protobuf which contains layout of all the sensors, so that
we don't waste bytes storing the names of sensors or how data is nested.
Without this protofile, google protobuf cannot really compress the
schemaless Node.proto, which can contain an arbitrary amount of nested
nodes. Without the schema, the serialized network table string is about
4000 bytes. With the schema, it's about 1000 bytes.

Also converted Tree.h to a bunch of helper functions instead of a class.