NAV-15 Connect NUC and BBB
This commit finishes up two projects:
nuc_eth_listener and bbb_eth_lister.
These run on either side of an ethernet cable
connecting the NUC and BBB, and are used
to send data between the two.

For example, sensor data and waypoints
are sent from the BBB to the NUC.
bbb_eth_listener is responsible
for sending this data, and nuc_eth_listener
is responsible for listening for it, and then
publishing it on the NUC side.

Right now, actuation angle (winch and motor angles)
are sent from the NUC to the BBB, so this data
will be passed along from nuc_eth_listener to bbb_eth_listener.

Note that communication between software on the NUC
is done via ROS publish/subscribe messages, so in order
to compile nuc_eth_listener, the cmake variable
ENABLE_ROS must be set to ON. It's in the top level
CMakeLists.txt