Sending simulator outputs to network table.
Most changes are in the nuc_to_bbb project.
Simulator outputs are received via a ROS publish message.
These are then sent directly to the computer. Right now
the code works when the network table is on the same device.
Code needs to be added to allow sending over an ethernet cable.

Note: By default, ROS is turned off, and the project
nuc_to_bbb will not be compiled. To enable it you have to
set the variable ENABLE_ROS to ON in top level CMakeLists.txt