NAV-12 Send sensor/uccm data to Rockblock
This is run on the BBB and sends Google Protobuf
serialized sensor and uccm diagnostic data from
the network table to the Rockblock.

The serial port path and the frequencies at which
the sensor and uccm data are sent are passed in as
arguments through the command line.

Note: Testing has only been done on a virtual
rockblock. The code has yet to be tested using
the real rockblock.