NAV 62 Send single protobuf type over satellite
Previously, the land server would receive protobuf
messages that could either be a Uccms.proto or a Sensors.proto,
and there was no way to tell which it was besides trying to parse
it as both. This code adds a new protobuf type called Satellite.proto,
which can hold either a Uccms.proto or a Sensors.proto, and has a flag
so that the land server can tell what type of data it is.
Technically this could have been done without the flag by using the
"has_uccms" and "has_sensors" fields, but I think the flag is cleaner.