NAV 69 Add/update READMEs
Many of the project folders were missing READMEs,
and information about how to run these programs
was passed around on slack. This information
is now available for lookup in various READMEs.

Other changes include adding the config_pins.sh script
which sets up canbus on the BBB, and updated the usage
string in bbb_rockblock_listener.