Fix bug where callback is never called. If the subscribe fails, then keep retrying it. The client needs to be guaranteed it has subscribed before continuing with the test. Sometimes the network server is down when we try to subscribe, which can cause the stress test to fail with error: callback is never called.
Fix bug where values_.txt would get wiped out if the server
crashed while writing to the file. Instead of writing directly to the
file, write to a swap file, delete the original file, and then
rename the swap file to the original. On startup,
just load the original file. If the original file is deleted
and the swap file exists, the swap file can be safely loaded,
since code that deletes the original file is after the code
which writes to the swap file.

Fix bug in test code where the test would do a GET and
get an old value. This was a race condition, where
the test would do a SET and then a GET, but in between
those two, another process would SET the value to
an older value (if it was farther behind in the code),
causing the equality check to fail.

Improve print statements in client. If the value was wrong
when doing a GET, print out the expected and actual values.

Various changes related to network_table_stress_test.
Changes made to the script itself, and changes
made to the network table and testing code in order to pass tests
on the BBB.

Optimization: When sending out subscribe replies, dont do
unneccessary serializations. Serialize the message once,
and then send it to everyone. This specific change
is in the function NotifySubscribers inside Server.cpp.

Allow network_table_stress_test.py to compile project if given
the --compile flag.

Add argument to network_table_stress_test.py to disable crashing.
use the --disable-fake-crashes flag.

Remove print statements from Connection.cpp