NAV-136
Fix corrupted subscriptions file.
The Network table uses this file to keep track of who is
subscribed, even after a crash. During the stress test,
this file was sometimes getting corrupted, causing
any subsequent instances of the network table to crash
when trying to parse the file.

This commit makes the network table
write to a temporary file first, then delete
the subscriptions file, and then rename the temp
file to the subscriptions file in order to prevent
it from being corrupted.

Also, even if a corrupt file is somehow opened, this
commit adds a try catch, where a corrupted file will
simply cause an error message to be printed out
and the file deleted, before the network table
continues to run.

Also made minor changes in Help.cpp (rename a variable
for consistency)