NAV-124
Jenkins can_bbb_nuc runs mock_sensors.py with the CSV file.
Before it was using real canbus data from can0, which was unpredictable.
Now it will playback sensor data on vcan0, so that tests are more
reproducable.

mock_sensors.py opens files relative to script location.
Or else we have to type out a absolute path in Jenkins, which is
annoying.

Updated pkill in Jenkinsfiles to use the -f flag.
This causes pkill to match the full command line, instead
of just the process name.

Mark failed stages as failed in Jenkinsfile.

Add a recorded csv file to scripts/recorded_data.