Run a cleanup before the build
If a previous test failed, it may cause leftover processes to still
be running, which can mess up subsequent tests.
Normally I would change the cleanup to always succeed, and just try
to clean up all the processes, so i don't need to do a cleanup at the
start of the build. However, because we're using daemonize and don't
have the outputs of the daemons, seeing which pkill failed is useful,
since it shows which process failed to run.

For example, if nuc_eth_listener fails to run and exits immediately with
an error message, you will see the pkill -f nuc_eth_listener is the last
pkill to run, which shows that nuc_eth_listener had a problem. If you
just made every pkill return true, you wouldnt be able to tell.

This is all hacky and should be gone when we stop using the daemonize
command. We should then be able to see the actual error messages when a
process crashes.