Memory leak fixes
Added code to handle SIGINT and gracefully terminate.
Without this, valgrind will report a memory leak when ctrl-c'ing the
server or connection.

Removed all calls to "new" in Server.cpp and Connection.cpp

The goal is to be able to use valgrind to test for memory leaks.
First step though is to remove all the current memory leaks.
This removes all leaks (detected by valgrind) in "network_table_server"
and "client" projects. I have not yet tested the other projects.