NAV-237: Global pathfinding startup in a new detached screen
All processes started in a detached screen (local-pathfinding, mock-ais, mock-sensors)
can be seen by running "screen -ls".
The screens can then be reattached to view the output of these processes.
ie. screen -r <SCREEN_ID>

Note that current global pathfinding command is not the one that will be run during launch.
The current command generates the waypoints a single time for a single destination, and does not connect to the network table.
Eventually we will need to replace this with the script that pathfinding creates which should continuously generate waypoints based on dynamic endpoints.