NAV-243: Implemented manual override interface
Added program that allows user input to manually control the rudder and winch.

A new rostopic and message type was added so that the NUC Ethernet Listener subscribes to both the regular and manual override topics. It prioritizes the manual override rostopic.

NUC Manual Override has a few options when running. It can operate continuously, prompting the user for input until explicitly told to exit. It can also also take a single input.

While the manual override is active, autonomous attempts to control the rudder and winch are blocked.

For full details on how to run the program, see projects/nuc_manual_override.