Corrected how winch and fib CAN frame were set.
Before, the angles for the winch and jib were interpreted
the same way as for the rudder angles. This would not have
worked as the rudder angles are floats while the winch and
jib angles are unsigned integers, so the reinterpret_cast<uint8_t *>
method used for the rudder angles would be wrong. It has been
changed to just use a static_cast<uint16_t>.