Nav 242 (#134)
* Large payloads can be parsed now.

Before, large payloads would fail to decode into utf-8. This is because
they are not encoded in utf-8, they're just binary. This was changed
and now the failure point is the uccm payload seems to have mismatched
attributes when using the stress_client on the BBB.

* Updated Help.py and uri.py

At some point the protobuf for UCCM was changed but the Python Help and
UCCM file were not. They are now updated and the land satellite listner
should work with payloads up to 340 bytes (maximum according to
rockblock) now.

* Fixed incorrect attribute in Help.py

Missed one error during last commit.

* Added basic test for future verification.

* Fixed linting

* Deleted temp error log